+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - HTTP
     - `ProxyServer` -- Fixed state of abruptly closed SSL interceptor connections
         leading to frozen browser operations.
+    - `Client::Dynamic404Handler`
+        - Added training scenario for when dashes are used as routing separators.
 - Session -- Allow for a submit input to be specified when the login needs to be
     triggered by clicking it, rather than just triggering the submit event on
     the form.